Delete journal.md
What did we do?
We have learned to program which we make dice with a graph underneath. 
What did we learn? 
 I have learned it was very difficult for me if I find the error. However, it was very fun whenever I change the error to collect coding. I got excited when the change was made and became perfect. I also really liked it when the dice was continuously going on and on. I felt the technology was very cool and fascinating. 
What question do I have?
At the end of the class, I still had an error so I really want to fix them with a teacher at the begging of the class.